Enable autocomplete for /yugipedia and /ygoprodeck, using got for HTTP/2 (#226)
Set up autocomplete framework and correct Australian spelling of "serialise"

Per https://github.com/dominictarr/bench-lru (and recent-enough verification in dominictarr/bench-lru#37), the best Map-compatible LRU cache is from mnemonist. We want to use an LRU cache to limit the memory use of these caches to be on the safe side, and it has to implement the Map API for use with keyv (internally used for got's HTTP cache). quick-lru, though suggested by keyv, is ESM-only now and the CJS fork @alloc/quick-lru is outdated. The Map implementation is chosen due to better performance with the string keys in use.

The httpCache follows HTTP caching semantics, so it should be working for Yugipedia, but not yet for YGOPRODECK since it is not supplying the headers. This speeds up subsequent autocompletes and cases where the user simply enters a value to search anyway.

The other cache stores the mapping of the canonical page or card names to links or identifiers that can be mapped to links. These are not expected to change often, and redirect anyway if changed. There is no expiry timer on these entries because fresh requests to the APIs where they are obtained would update these anyway, except where a name might be no longer valid.

Closes #197
Closes #198